Improve error message when receiving unexpected record content via Broker.
Broker sends record values as a vector of the corresponding fields. If
the received values doesn't match the actual record fields, Zeek has
so far been reporting this not-so-helpful error message:

    warning: failed to convert remote event 'got_result' arg #1,  got vector, expected record

This change turns the latter part into a list of the field types we
actually received, which makes debugging much easier.

(No test, seems overkill to add a spawn a Broker client to exercise
this.)